fix input and resolve type issues
