+++ packages/ui/src/lib/input/types.ts
@@ -1,6 +1,32 @@
 import type { Snippet } from 'svelte';
-import type { ChangeEventHandler } from 'svelte/elements';
 
+import type {
+	ChangeEventHandler,
+	HTMLInputAttributes,
+	HTMLInputTypeAttribute,
+	HTMLTextareaAttributes
+} from 'svelte/elements';
+
+export type FormatFunction = (
+	value: string,
+	details?: {
+		name?: string;
+		type?: string;
+		disabled?: boolean;
+	}
+) => string | number;
+
+export type InputMode =
+	| 'none'
+	| 'search'
+	| 'text'
+	| 'tel'
+	| 'url'
+	| 'email'
+	| 'numeric'
+	| 'decimal'
+	| null
+	| undefined;
 export interface InputProps {
 	/**
 	 * String that labels the input.
@@ -65,3 +91,52 @@ export interface InputProps {
 	children?: Snippet;
 	onchange?: ChangeEventHandler<HTMLInputElement>;
 }
+
+interface InputPropsBase {
+
+	/**
+	 * The value of the input. Can be bound to and externally modified.
+	 */
+	value?: string;
+
+	/**
+	 * The `inputmode` of the `<input>` element, which provides a hint about what type of virtual keyboard to display.
+	 */
+	inputmode?: InputMode;
+
+	/**
+	 * Text that appears within the `<input>` element when no value is present.
+	 */
+	placeholder?: string;
+
+	/**
+	 * Function that will be applied to transform the value when the input element loses focus.
+	 * By default, it trims leading and trailing whitespace (but does nothing if `type` is `password`).
+	 */
+	format: null | FormatFunction;
+}
+
+export interface InputAsTextArea extends InputPropsBase, Omit<InputProps, 'onchange'> {
+	/**
+	 * The `type` of the `<input>` element (see [MDN docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#input_types)).
+	 *
+	 * Additionally, passing `textarea` will render a `<textarea>` instead of an `<input>`.
+	 */
+
+	type: 'textarea',
+	restProps: HTMLTextareaAttributes;
+	onchange: ChangeEventHandler<HTMLTextAreaElement>;
+}
+
+export interface InputAsNonTextArea extends InputPropsBase, InputProps {
+	/**
+	 * The `type` of the `<input>` element (see [MDN docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#input_types)).
+	 *
+	 * Additionally, passing `textarea` will render a `<textarea>` instead of an `<input>`.
+	 */
+	type: Exclude<HTMLInputTypeAttribute, 'textarea'>,
+	restProps: HTMLInputAttributes;
+}
+
+export type Props = InputAsTextArea | InputAsNonTextArea;
+
